resume-matcher-v2/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pdfminer.six==20240706
PyPDF2==3.0.1
python-docx==1.1.2
docx2txt==0.8
scikit-learn==1.5.2
numpy==2.1.3
scipy==1.13.1
orjson==3.10.7
# pyarrow==0
pyarrow==16.1.0

